fix(auth): protect customer supplier and category writes

+++ pc-api/routes/api/customer.php
@@ -13,30 +13,30 @@ Route::prefix('customers')->middleware(['auth:sanctum', 'ensure_business', 'perm
     Route::get('stats', [CustomerController::class, 'stats']);
     Route::get('industries', [CustomerController::class, 'industries']);
     Route::get('health', [CustomerController::class, 'health']);
-    Route::post('import', [CustomerController::class, 'import']);
+    Route::post('import', [CustomerController::class, 'import'])->middleware('permission:customer.create');
     // 销售漏斗看板
     Route::get('pipeline', [CustomerPipelineController::class, 'index']);
     Route::get('pipeline/weekly-trend', [CustomerPipelineController::class, 'weeklyTrend']);
     Route::get('map', [CustomerController::class, 'mapData']);
     Route::get('{customer}/profile', [CustomerController::class, 'profile']);
     Route::get('{customer}/follow-ups', [CustomerController::class, 'followUps']);
-    Route::post('{customer}/follow-ups', [CustomerController::class, 'storeFollowUp']);
+    Route::post('{customer}/follow-ups', [CustomerController::class, 'storeFollowUp'])->middleware('permission:customer.edit');
     Route::get('{customer}/devices', [CustomerController::class, 'devices']);
     // 联系人管理
     Route::get('{customer}/contacts', [CustomerController::class, 'listContacts']);
-    Route::post('{customer}/contacts', [CustomerController::class, 'storeContact']);
-    Route::put('{customer}/contacts/{contact}', [CustomerController::class, 'updateContact'])->whereNumber('contact');
-    Route::delete('{customer}/contacts/{contact}', [CustomerController::class, 'destroyContact'])->whereNumber('contact');
+    Route::post('{customer}/contacts', [CustomerController::class, 'storeContact'])->middleware('permission:customer.edit');
+    Route::put('{customer}/contacts/{contact}', [CustomerController::class, 'updateContact'])->whereNumber('contact')->middleware('permission:customer.edit');
+    Route::delete('{customer}/contacts/{contact}', [CustomerController::class, 'destroyContact'])->whereNumber('contact')->middleware('permission:customer.edit');
     // 开票信息
     Route::get('{customer}/invoice-infos', [CustomerController::class, 'listInvoiceInfos']);
-    Route::post('{customer}/invoice-infos', [CustomerController::class, 'storeInvoiceInfo']);
-    Route::put('{customer}/invoice-infos/{info}', [CustomerController::class, 'updateInvoiceInfo'])->whereNumber('info');
-    Route::delete('{customer}/invoice-infos/{info}', [CustomerController::class, 'destroyInvoiceInfo'])->whereNumber('info');
+    Route::post('{customer}/invoice-infos', [CustomerController::class, 'storeInvoiceInfo'])->middleware('permission:customer.edit');
+    Route::put('{customer}/invoice-infos/{info}', [CustomerController::class, 'updateInvoiceInfo'])->whereNumber('info')->middleware('permission:customer.edit');
+    Route::delete('{customer}/invoice-infos/{info}', [CustomerController::class, 'destroyInvoiceInfo'])->whereNumber('info')->middleware('permission:customer.edit');
     // 放最后
     Route::get('{customer}', [CustomerController::class, 'show']);
-    Route::put('{customer}', [CustomerController::class, 'update']);
-    Route::put('{customer}/stage', [CustomerPipelineController::class, 'updateStage']);
-    Route::delete('{customer}', [CustomerController::class, 'destroy']);
+    Route::put('{customer}', [CustomerController::class, 'update'])->middleware('permission:customer.edit');
+    Route::put('{customer}/stage', [CustomerPipelineController::class, 'updateStage'])->middleware('permission:customer.edit');
+    Route::delete('{customer}', [CustomerController::class, 'destroy'])->middleware('permission:customer.edit');
 });
 
 // ========== 供应商管理 ==========
@@ -44,10 +44,10 @@ Route::prefix('suppliers')->middleware(['auth:sanctum', 'ensure_business', 'perm
     Route::get('/', [SupplierController::class, 'index']);
     Route::post('/', [SupplierController::class, 'store'])->middleware('permission:supplier.create');
     Route::get('{id}', [SupplierController::class, 'show'])->whereNumber('id');
-    Route::put('{id}', [SupplierController::class, 'update'])->whereNumber('id');
-    Route::delete('{id}', [SupplierController::class, 'destroy'])->whereNumber('id');
-    Route::post('{id}/change-status', [SupplierController::class, 'changeStatus'])->whereNumber('id');
-    Route::post('{id}/sync-contacts', [SupplierController::class, 'syncContacts'])->whereNumber('id');
+    Route::put('{id}', [SupplierController::class, 'update'])->whereNumber('id')->middleware('permission:supplier.create');
+    Route::delete('{id}', [SupplierController::class, 'destroy'])->whereNumber('id')->middleware('permission:supplier.create');
+    Route::post('{id}/change-status', [SupplierController::class, 'changeStatus'])->whereNumber('id')->middleware('permission:supplier.create');
+    Route::post('{id}/sync-contacts', [SupplierController::class, 'syncContacts'])->whereNumber('id')->middleware('permission:supplier.create');
     Route::get('{id}/evaluations', [SupplierController::class, 'evaluations'])->whereNumber('id');
 });
